LayoutTests/svg/custom/html_document_set_title.html

<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
  function runTest()
  {
    shouldBe('document.title', '""');
  }
</script>
</head>
<body onload="runTest()">
<svg>
<title>Should not be set on HTML Document</title>
</svg>
</body>
</html>
